+++ src/main/kotlin/com/fobgochod/git/commit/action/RestoreTypesAction.kt
@@ -1,14 +1,12 @@
 package com.fobgochod.git.commit.action
 
 import com.fobgochod.git.commit.domain.CommitType
-import com.fobgochod.git.commit.domain.TypeRow
 import com.fobgochod.git.commit.domain.TypeTable
 import com.fobgochod.git.commit.util.GitBundle
 import com.fobgochod.git.commit.util.GitIcons
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.ui.AnActionButton
-import java.util.*
 
 class RestoreTypesAction(private val typeTable: TypeTable) : AnActionButton(
     GitBundle.message("action.toolbar.restore.text"),
@@ -16,23 +14,13 @@ class RestoreTypesAction(private val typeTable: TypeTable) : AnActionButton(
     GitIcons.RESTORE_ACTION
 ) {
 
-    private val typeRows: MutableList<TypeRow> = LinkedList()
-
-    init {
-        if (typeRows.isEmpty()) {
-            for (type in CommitType.values()) {
-                typeRows.add(TypeRow(type.type(), type.description()))
-            }
-        }
-    }
-
     override fun actionPerformed(event: AnActionEvent) {
-        typeTable.reset(typeRows)
+        typeTable.reset(CommitType.typeRows)
     }
 
     override fun updateButton(event: AnActionEvent) {
         // super.update(event)
-        event.presentation.isEnabled = typeRows != typeTable.typeRows
+        event.presentation.isEnabled = CommitType.typeRows != typeTable.typeRows
     }
 
     override fun getActionUpdateThread(): ActionUpdateThread {