+++ src/main/kotlin/com/fobgochod/git/commit/constant/GitConstant.kt
@@ -13,8 +13,7 @@ object GitConstant {
 
     const val BREAKING_CHANGE_PREFIX = "BREAKING CHANGE: "
     const val CLOSES_PREFIX = "Closes "
-    const val SKIP_CI = "[skip ci]"
 
-    val HEADER_PATTERN: Pattern = Pattern.compile("^([a-z]+)(\\((.+)\\))?: (.+)")
+    val HEADER_PATTERN: Pattern = Pattern.compile("^([a-z]+)(\\((.+)?\\))?: (.+)?")
     val CLOSED_ISSUES_PATTERN: Pattern = Pattern.compile("Closes (.+)")
 }