+++ src/main/kotlin/com/fobgochod/git/commit/settings/GitConfigurable.kt
@@ -1,67 +1,33 @@
 package com.fobgochod.git.commit.settings
 
-import com.fobgochod.git.commit.domain.ViewForm
 import com.fobgochod.git.commit.util.GitBundle
-import com.intellij.openapi.options.SearchableConfigurable
-import org.jetbrains.annotations.Nls
-import javax.swing.JComponent
+import com.intellij.openapi.options.BoundSearchableConfigurable
+import com.intellij.openapi.ui.DialogPanel
 
-class GitConfigurable : SearchableConfigurable {
+class GitConfigurable :
+    BoundSearchableConfigurable(GitBundle.message("configurable.display.name"), GitConfigurable::class.java.name) {
 
-    private var state: GitState = GitState.getInstance()
-    private var component: GitComponent = GitComponent()
+    private var myPanel: GitPanel = GitPanel()
 
     override fun getId(): String {
         return GitConfigurable::class.java.name
     }
 
-    override fun createComponent(): JComponent {
-        return component.mainPanel
-    }
-
-    override fun getDisplayName(): @Nls(capitalization = Nls.Capitalization.Title) String {
-        return GitBundle.message("plugin.name")
+    override fun createPanel(): DialogPanel {
+        return myPanel.createPanel()
     }
 
     override fun reset() {
-        component.typeTable.reset(state.typeRows)
-        component.typeCountField.text = state.typeCount.toString()
-        component.viewMode.selectedItem = state.viewMode
-
-        for (checkBox in component.viewForm) {
-            val viewForm = ViewForm.valueOf(checkBox.text)
-            checkBox.isSelected = state.isViewFormHidden(viewForm)
-        }
+        super.reset()
+        myPanel.reset()
     }
 
     override fun isModified(): Boolean {
-        val typeRows = component.typeTable.typeRows
-        val typeCount: Int = component.typeCountField.text.toInt()
-        val viewMode = component.getViewMode()
-        return typeRows != state.typeRows
-                || typeCount != state.typeCount
-                || viewMode != state.viewMode
-                || isViewFormModified()
-    }
-
-    private fun isViewFormModified(): Boolean {
-        for (checkBox in component.viewForm) {
-            val viewForm = ViewForm.valueOf(checkBox.text)
-            if (state.isViewFormHidden(viewForm) != checkBox.isSelected) {
-                return true
-            }
-        }
-        return false
+        return super.isModified() || myPanel.isModified()
     }
 
     override fun apply() {
-        state.typeRows = component.typeTable.typeRows
-        state.typeCount = component.typeCountField.text.toInt()
-        state.viewMode = component.getViewMode()
-
-        for (checkBox in component.viewForm) {
-            val formItem = ViewForm.valueOf(checkBox.text)
-            state.viewForm[formItem] = checkBox.isSelected
-        }
+        super.apply()
+        myPanel.apply()
     }
 }